Fix playlist

+++ app/module/playlist/src/entity/mixin/DurationMixin.js
@@ -19,7 +19,7 @@ export const DurationMixin = (superClass) => {
 
             let duration = 0;
             for (let cont = 0; this.resources.length > cont; cont++) {
-                duration = duration + parseInt(this.resources[cont].duration);
+                duration = duration + parseFloat(this.resources[cont].duration);
             }
             return duration;
         }
